Length of Wood and number of slots
A large improvement in code refactoring, comments and clarity.

Also added the ability to specify the length of the wood to be joined along with a number of slots.  This will determine the size of each slot which should make it easier to get a nicely finished result when the size of the wood is fixed.